chore(internal/config): move SpanStartSnapshot to snapshots.go
Separate file for the snapshot pattern. config.go is dominated by
per-field getter/setter pairs; snapshots are conceptually distinct and
likely to grow as more hot paths adopt the pattern. Top-of-file comment
establishes the per-caller convention so future snapshots have an
obvious home.
